banshee: Broken first attempt at DMA impl
A first attempt at getting the DMA to move data. The implemntation is
currently broken, failing with a null pointer access during JIT
execution.

To fix this, we'll have to split the runtime into a shared part which is
compiled into banshee's Rust code, and a JIT-only part which is compiled
into the `runtime.ll` file. Also, the `Cpu` and `CpuState` structs need
to move over into the shared part of the runtime.