banshee: Make banshee/JIT split for runtime clearer
Clean the runtime up quite a bit. Now we have two groups of files in
src/runtime: jit.{ll,rs} which are linked into the JITed binary, and
mod.rs which is compiled into banshee. A common.rs file is there to hold
code which is common to both.

The build script now automatically calls rustc in order to compile the
jit.rs file into its corresponding LLVM IR.