banshee: Fix nasty dominance issue in SSR tracing
Fix a nasty bug where instructions that perform SSR accesses can lead
to instructions in the IR not dominating all their uses when instruction
tracing is enabled. This happens because the translator allocates a
fixed slot for the SSR memory/register accesses, but the corresponding
value is only produced conditionally if SSRs are enabled. Tracing
happens unconditionally though, so the tracer tries to read a value that
does not dominate it.

The long-term fix will be to make instruction tracing dynamic, with a
function in `jit.rs` that tracks up to 8 (or any buffer size) accesses
in a local tracing buffer. Then we can emit dynamic tracing for SSR
accesses as well.

For now, SSR access tracing is just disabled.