banshee: Fix overflow issue with long branches
The branch immediate encoded in branch instructions is 12 bits wide, but
since is implicitly a 2-byte offset. So for real target computation, the
immediate is shifted one bit to the left. The rust decoder generated by
`parse-opcodes` truncates to 12 bits and sign extends *after* the shift
by one. This essentially discards the topmost bit of the branch
immediate.

This fixes the issue and adds a corresponding test.